ObjC: ARC migration, autorelease pools, and canonical error ids
Migrate the Metal bridge to ARC with a #error guard, wrap every cgo entry
point in @autoreleasepool, switch buffer_new to size_t, detect non-uniform
threadgroup support once at init for the dispatch path, and make function_run
report "invalid function id"/"invalid buffer id" so the Go sentinels match.

+++ Metal.h
@@ -18,7 +18,7 @@ const char *function_name(int functionId);
 
 // Functions that must be called once for every buffer used as an argument to
 // a metal function
-int buffer_new(int size, void **contents, const char **error);
+int buffer_new(size_t size, void **contents, const char **error);
 _Bool buffer_close(int bufferId, const char **error);
 
 // Functions for closing metal resources
